+++ core/amber/src/main/python/core/architecture/handlers/control/add_input_channel_handler.py
@@ -7,7 +7,7 @@ class AddInputChannelHandler(ControlHandler):
     cmd = AddInputChannelV2
 
     def __call__(self, context: Context, command: AddInputChannelV2, *args, **kwargs):
-        context.batch_to_tuple_converter.register_input(
+        context.input_manager.register_input(
             command.channel_id, command.port_id
         )
         return None
