+++ core/amber/src/main/python/core/architecture/handlers/control/add_partitioning_handler.py
@@ -7,7 +7,7 @@ class AddPartitioningHandler(ControlHandler):
     cmd = AddPartitioningV2
 
     def __call__(self, context: Context, command: AddPartitioningV2, *args, **kwargs):
-        context.tuple_to_batch_converter.add_partitioning(
+        context.output_manager.add_partitioning(
             command.tag, command.partitioning
         )
         return None
