+++ core/amber/src/main/python/core/architecture/handlers/control/start_worker_handler.py
@@ -4,7 +4,7 @@ from proto.edu.uci.ics.amber.engine.architecture.worker import (
 )
 from core.architecture.handlers.control.control_handler_base import ControlHandler
 from core.architecture.managers.context import Context
-from core.architecture.packaging.batch_to_tuple_converter import BatchToTupleConverter
+from core.architecture.packaging.input_manager import InputManager
 from core.models.internal_queue import DataElement
 
 from loguru import logger
@@ -18,7 +18,7 @@ class StartWorkerHandler(ControlHandler):
         if context.operator_manager.operator.is_source:
             context.state_manager.transit_to(WorkerState.RUNNING)
             context.input_queue.put(
-                DataElement(tag=BatchToTupleConverter.SOURCE_STARTER, payload=None)
+                DataElement(tag=InputManager.SOURCE_STARTER, payload=None)
             )
         state = context.state_manager.get_current_state()
         return state
