+++ core/amber/src/main/python/core/runnables/main_loop.py
@@ -10,7 +10,7 @@ from pampy import match
 
 from core.architecture.managers.context import Context
 from core.architecture.managers.pause_manager import PauseType
-from core.architecture.packaging.batch_to_tuple_converter import EndOfAllMarker
+from core.architecture.packaging.input_manager import EndOfAllMarker
 from core.architecture.rpc.async_rpc_client import AsyncRPCClient
 from core.architecture.rpc.async_rpc_server import AsyncRPCServer
 from core.models import (
@@ -165,7 +165,7 @@ class MainLoop(StoppableQueueBlockingRunnable):
                 for (
                     to,
                     batch,
-                ) in self.context.tuple_to_batch_converter.tuple_to_batch(output_tuple):
+                ) in self.context.output_manager.tuple_to_batch(output_tuple):
                     batch.schema = self.context.operator_manager.operator.output_schema
                     self._output_queue.put(DataElement(tag=to, payload=batch))
 
@@ -222,7 +222,7 @@ class MainLoop(StoppableQueueBlockingRunnable):
         :param sender_change_marker: SenderChangeMarker which contains sender link.
         """
         self.context.tuple_processing_manager.current_input_port_id = (
-            self.context.batch_to_tuple_converter.get_port_id(
+            self.context.input_manager.get_port_id(
                 sender_change_marker.channel_id
             )
         )
@@ -236,7 +236,7 @@ class MainLoop(StoppableQueueBlockingRunnable):
 
         :param _: EndOfAllMarker
         """
-        for to, batch in self.context.tuple_to_batch_converter.emit_end_of_upstream():
+        for to, batch in self.context.output_manager.emit_end_of_upstream():
             batch.schema = self.context.operator_manager.operator.output_schema
             self._output_queue.put(DataElement(tag=to, payload=batch))
             self._check_and_process_control()
@@ -261,7 +261,7 @@ class MainLoop(StoppableQueueBlockingRunnable):
             self.context.state_manager.transit_to(WorkerState.RUNNING)
 
         self.context.tuple_processing_manager.current_input_tuple_iter = (
-            self.context.batch_to_tuple_converter.process_data_payload(
+            self.context.input_manager.process_data_payload(
                 data_element.tag, data_element.payload
             )
         )
